Add locking and proper error reporting for permissions
Now two jumpstarter instances can't access the same port concurrently,
performing commands on a busy device will return an error, and
the busy devices are listed in BUSY state.

If you try to open a serial port and you don't have permissions for
the device an error will be reported too.